refactor: [#76] extract catalog module
